+++ mkl_fft/_pydfti.pyx
@@ -89,7 +89,7 @@ def _tls_dfti_cache_capsule():
 cdef extern from "Python.h":
     ctypedef int size_t
 
-    long PyInt_AsLong(object ob)
+    long PyLong_AsLong(object ob)
     int PyObject_HasAttrString(object, char*)
 
 
@@ -262,7 +262,7 @@ cdef cnp.ndarray _process_arguments(
     xnd[0] = cnp.PyArray_NDIM(x_arr)  # tensor-rank of the array
 
     err = 0
-    axis_[0] = PyInt_AsLong(axis)
+    axis_[0] = PyLong_AsLong(axis)
     if (axis_[0] == -1 and PyErr_Occurred()):
         PyErr_Clear()
         err = 1
@@ -278,7 +278,7 @@ cdef cnp.ndarray _process_arguments(
         n_[0] = x_arr.shape[axis_[0]]
     else:
         try:
-            n_[0] = PyInt_AsLong(n)
+            n_[0] = PyLong_AsLong(n)
         except:
             err = 1
 
@@ -334,7 +334,7 @@ cdef int _is_integral(object num):
     if num is None:
         return 0
     try:
-        n = PyInt_AsLong(num)
+        n = PyLong_AsLong(num)
         _integral = 1 if n > 0 else 0
     except:
         _integral = 0
